finished updating the namespace information
